Django/django_api_react/requirements.txt
amqp==5.0.9
asgiref==3.4.1
asttokens==2.0.5
async-timeout==4.0.2
attrs==21.2.0
autopep8==1.6.0
backcall==0.2.0
backoff==1.11.1
backports.entry-points-selectable==1.1.0
beautifulsoup4==4.9.3
billiard==3.6.4.0
bson==0.5.10
celery==5.2.3
certifi==2021.10.8
cffi==1.15.0
chardet==4.0.0
charset-normalizer==2.0.7
click==8.0.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.4
cryptography==36.0.1
debugpy==1.6.0
decorator==5.1.0
defusedxml==0.7.1
Deprecated==1.2.13
distlib==0.3.3
Django==3.2.9
django-allauth==0.48.0
django-cors-headers==3.10.0
django-debug-toolbar==3.2.1
django-redis==5.2.0
django-silk==4.2.0
djangorestframework==3.12.4
djangorestframework-simplejwt==5.0.0
dnspython==2.0.0
drf-spectacular==0.21.0
ecdsa==0.17.0
entrypoints==0.4
executing==0.8.3
filelock==3.4.0
gprof2dot==2021.2.21
gunicorn==20.1.0
httpretty==1.1.4
idna==3.3
importlib-resources==5.4.0
inflection==0.5.1
ipykernel==6.13.0
ipython==8.3.0
jedi==0.18.1
Jinja2==3.1.1
joblib==1.0.1
jsonschema==4.2.1
jupyter-client==7.3.1
jupyter-core==4.10.0
kombu==5.2.3
MarkupSafe==2.1.1
matplotlib-inline==0.1.3
nest-asyncio==1.5.5
oauthlib==3.2.0
packaging==21.3
parso==0.8.3
pickleshare==0.7.5
Pillow==8.4.0
platformdirs==2.4.0
postgres==4.0
prompt-toolkit==3.0.24
psutil==5.9.0
psycopg2==2.9.1
psycopg2-binary==2.8.6
psycopg2-pool==1.1
pure-eval==0.2.2
pyasn1==0.4.8
pycodestyle==2.8.0
pycparser==2.21
Pygments==2.12.0
PyJWT==2.3.0
pyOpenSSL==21.0.0
pyparsing==3.0.9
pyrsistent==0.18.0
python-dateutil==2.8.2
python-dotenv==0.19.2
python-jose==3.3.0
python-keycloak-client==0.2.3
python3-openid==3.2.0
pytz==2021.3
pywin32==305
PyYAML==6.0
pyzmq==22.3.0
redis==4.3.4
regex==2021.8.3
requests==2.26.0
requests-file==1.5.1
requests-oauthlib==1.3.1
rsa==4.8
sentry-sdk==1.5.12
six==1.16.0
soupsieve==2.2.1
sqlparse==0.4.2
stack-data==0.2.0
tldextract==3.2.0
toml==0.10.2
tornado==6.1
tqdm==4.62.1
traitlets==5.2.0
uritemplate==4.1.1
urllib3==1.26.7
validators==0.18.2
vine==5.0.0
virtualenv==20.8.1
wcwidth==0.2.5
whitenoise==5.3.0
wrapt==1.14.1
zipp==3.6.0
